Add URL Encoding and Decoding Functionality
Added new URL encoding and decoding functionality in `urlencode.ts` to replace spaces with "+" instead of "%20". This update introduces four new functions (`encodeFormElement`, `decodeFormElement`, `encodeFormData`, `decodeFormData`) to properly encode and decode string for URL. The legacy `urlencode` and `urldecode` functions are now marked as deprecated and they should be replaced with `encodeFormData` and `decodeFormData` respectively.